ref: Move ratelimiter to its own file
